python/ProcessEngine/pyproject.toml
[project]
name = "processengine"
version = "0.1.0"
description = "Greenglass Project Python Process Engine"
authors = [
    { name = "Steve", email = "steve.hopkins@kathris.com" }
]
dependencies = [
    "asyncio>=3.4.3",
    "websockets>=12.0",
    "pyyaml>=6.0.1",
    "rx>=3.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/processengine"]
